% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mean.R
\name{cluster_mean}
\alias{cluster_mean}
\title{Perform cluster mean centering}
\usage{
cluster_mean(df, x, g, label = x, clobber = FALSE)
}
\arguments{
\item{df}{data.frame containing the variables}

\item{x}{char; label of the column to cluster mean center}

\item{g}{char; label of the column with cluster ids}

\item{label}{char; to use for new columns (default=x)}

\item{clobber}{logical; to control behavior when a column <label>_m or <label>_s already exists. TRUE: overwrite existing column. FALSE: exit function (default=FALSE)}
}
\value{
data.frame,tibble
}
\description{
For the given data frame,
calculate the mean of the given column ('x')
within each cluster defined by column 'g'.

A new data.frame is returned with two additional columns,
  *_m the mean of x within each cluster
  *_s the individual values of x, centered at 0
}
\examples{
df <- tibble(x=rnorm(10), g=rep(c(0,1),5))
df <- cluster_mean(df, 'x', 'g')
df
# A tibble: 10 x 4
          x     g    x_m     x_s
      <dbl> <dbl>  <dbl>   <dbl>
 1 -0.918       0 -0.565 -0.353
 2  0.355       1  0.280  0.0753
 3  0.0313      0 -0.565  0.596
 4  0.858       1  0.280  0.578
 5 -0.201       0 -0.565  0.364
 6  1.88        1  0.280  1.60
 7 -0.907       0 -0.565 -0.342
 8 -1.69        1  0.280 -1.97
 9 -0.830       0 -0.565 -0.265
10 -0.00634     1  0.280 -0.286

}
