% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_scale.R
\docType{package}
\name{cluster_scale}
\alias{cluster_scale}
\title{Perform cluster mean centering and convert to cluster-specific z-scores}
\usage{
cluster_scale(df, x, g, label, clobber = FALSE)
}
\arguments{
\item{df}{data.frame containing the variables}

\item{x}{char; label of the column to cluster mean center}

\item{g}{char; label of the column with cluster ids}

\item{label}{char; to use for new columns (default=x)}

\item{clobber}{logical; to control behavior when a column <label>_m or <label>_z already exists. TRUE: overwrite existing column. FALSE: exit function (default=FALSE)}
}
\value{
data.frame,tibble
}
\description{
For the given data frame,
calculate the mean of the given column ('x')
within each cluster defined by column 'g'.

A new data.frame is returned with two additional columns,
  *_m the mean of x within each cluster
  *_z the z-scored values of x
}
\examples{
df <- tibble(x=rnorm(10), g=rep(c(0,1),5))
df <- cluster_scale(df, x, g)
df
# A tibble: 10 x 4
      x     g   x_m    x_z
  <dbl> <dbl> <dbl>  <dbl>
1 -0.680     0 0.260 -1.10
2  0.927     1 0.350  0.484
3 -0.590     0 0.260 -0.993
4  1.44      1 0.350  0.912
5  1.10      0 0.260  0.984
6 -0.625     1 0.350 -0.817
7  0.447     0 0.260  0.218
8 -1.22      1 0.350 -1.31
9  1.02      0 0.260  0.889
10  1.23      1 0.350  0.735

}
