% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_ortho.R
\docType{package}
\name{poly_ortho}
\alias{poly_ortho}
\title{Calculate and return the i-th ordered orthogonal polynomial}
\usage{
poly_ortho(x, ord = 2)
}
\arguments{
\item{x}{numeric: data from which to calculate polynomials}

\item{ord}{int: order of polynomial to extract (0 < ord < 5)}
}
\value{
numeric
}
\description{
For the given vector, return the orthogonal
parameterization of the i-th ordered polynomial
}
\details{
poly_ortho uses stats::poly to calculate orthogonal polynomials
of the given data. These are necessarily to examine these
non-linear effects in models while avoiding co-linearity
between lower order polynomials.

The returned orthogonalized vectors may be useful for creating
design matrices in contexts were polynomials are desired, but the
poly function cannot be easily used (e.g. with rstan functions)
}
\examples{
x <- sort(rnorm(5))
poly_ortho(x, 1)
[1] -0.45049407 -0.25381561 -0.09458044 -0.05031853  0.84920864
poly_ortho(x, 2)
[1] -0.34934124  0.81629373 -0.03666156 -0.45774186  0.02745094
poly_ortho(x, 3)
[1]  0.68735315 -0.02314875 -0.42889989 -0.51474832  0.27944380
}
\seealso{
\code{\link[stats]{poly}}
}
